+++ packages/setup-ocaml/src/constants.ts
@@ -42,9 +42,9 @@ export const CYGWIN_ROOT_BIN = path.join(CYGWIN_ROOT, "bin");
 export const CYGWIN_ROOT_WRAPPERBIN = path.join(CYGWIN_ROOT, "wrapperbin");
 
 // [todo] remove the branch for Windows once opam 2.2 is released as stable.
-export const ALLOW_PRELEASE_OPAM =
+export const ALLOW_PRERELEASE_OPAM =
   PLATFORM !== "win32" &&
-  core.getBooleanInput("allow-prelease-opam", {
+  core.getBooleanInput("allow-prerelease-opam", {
     required: false,
     trimWhitespace: true,
   });
@@ -71,7 +71,7 @@ export const OCAML_COMPILER = core.getInput("ocaml-compiler", {
 
 // [todo] remove this once opam 2.2 is released as stable.
 export const OPAM_DEPEXT =
-  !ALLOW_PRELEASE_OPAM &&
+  !ALLOW_PRERELEASE_OPAM &&
   core.getBooleanInput("opam-depext", {
     required: false,
     trimWhitespace: true,